Keep antd and react-dom out of the bucket-data web workers
async_get_maximum_zoom_for_all_mags.worker was 703 kB, of which ~350 kB was
main-thread UI code that a pure-math worker has no use for: react-dom (172 kB),
antd, @ant-design/cssinjs, stylis, fast-color, @airbrake/browser and cross-fetch.

Two static import edges pulled it in, both reached via flycam_accessor:

  worker -> flycam_accessor -> data_rendering_logic -> libs/toast -> antd
  worker -> flycam_accessor -> dataset_accessor    -> libs/error_handling -> libs/toast

Neither module needs the UI on the worker path:

- data_rendering_logic's only Toast/ErrorHandling users were getSupportedTextureSpecs,
  guardAgainstMesaLimit and validateMinimumRequirements. Those probe a WebGL context and
  need `document`, so they can only ever run on the main thread. Moved them to a new
  gpu_capability_check module, leaving data_rendering_logic as pure computation.

- dataset_accessor used ErrorHandling.assert once, in getLayerByName. Added libs/assertion,
  a tiny worker-safe assert whose reporter libs/error_handling installs on load, so the
  main thread still gets the toast, the Airbrake report and throwAssertions, while workers
  fall back to console.warn.

Worker drops from 702.75 kB to 350.35 kB; the 218 kB of three that remains is the
matrix math the worker actually runs.
