+++ frontend/javascripts/viewer/model/bucket_data_handling/data_rendering_logic.tsx
@@ -1,6 +1,3 @@
-import ErrorHandling from "libs/error_handling";
-import Toast from "libs/toast";
-import { document } from "libs/window";
 import max from "lodash-es/max";
 import memoize from "lodash-es/memoize";
 import min from "lodash-es/min";
@@ -20,111 +17,11 @@ import type { ElementClass } from "types/api_types";
 import constants from "viewer/constants";
 import type { TypedArrayConstructor } from "../helpers/typed_buffer";
 
-type GpuSpecs = {
+export type GpuSpecs = {
   supportedTextureSize: number;
   maxTextureCount: number;
 };
 const lookupTextureCount = 1;
-export function getSupportedTextureSpecs(): GpuSpecs {
-  const canvas = document.createElement("canvas");
-  const contextProvider =
-    "getContext" in canvas
-      ? (ctxName: "webgl2") => canvas.getContext(ctxName)
-      : (ctxName: string) => ({
-          MAX_TEXTURE_SIZE: 0,
-          MAX_TEXTURE_IMAGE_UNITS: 1,
-
-          getParameter(param: number) {
-            if (ctxName === "webgl2") {
-              const dummyValues: Record<string, any> = {
-                "0": 4096,
-                "1": 16,
-                "4": "debugInfo.UNMASKED_RENDERER_WEBGL",
-                "7937": "Radeon R9 200 Series",
-              };
-              return dummyValues[param];
-            }
-
-            throw new Error(`Unknown call to getParameter: ${param}`);
-          },
-
-          getExtension(param: string) {
-            if (param === "WEBGL_debug_renderer_info") {
-              return {
-                UNMASKED_RENDERER_WEBGL: 4,
-              };
-            }
-
-            throw new Error(`Unknown call to getExtension: ${param}`);
-          },
-        });
-  const gl = contextProvider("webgl2");
-
-  if (!gl) {
-    Toast.error(
-      <span>
-        Your browser does not seem to support WebGL 2. Please upgrade your browser or hardware and
-        ensure that WebGL 2 is supported. You might want to use{" "}
-        <a href="https://get.webgl.org/webgl2/" target="_blank" rel="noreferrer">
-          this site
-        </a>{" "}
-        to check the WebGL support yourself.
-      </span>,
-      {
-        sticky: true,
-      },
-    );
-    throw new Error("WebGL2 context could not be constructed.");
-  }
-
-  const supportedTextureSize = gl.getParameter(gl.MAX_TEXTURE_SIZE);
-  const maxTextureImageUnits = gl.getParameter(gl.MAX_TEXTURE_IMAGE_UNITS);
-
-  if (import.meta.env.MODE !== "test") {
-    console.log("maxTextureImageUnits", maxTextureImageUnits);
-  }
-
-  return {
-    supportedTextureSize,
-    maxTextureCount: guardAgainstMesaLimit(maxTextureImageUnits, gl),
-  };
-}
-
-function guardAgainstMesaLimit(maxSamplers: number, gl: any) {
-  // Adapted from here: https://github.com/pixijs/pixi.js/pull/6354/files
-
-  try {
-    let renderer = gl.getParameter(gl.RENDERER);
-    if (renderer == null) {
-      const debugInfo = gl.getExtension("WEBGL_debug_renderer_info");
-
-      if (debugInfo != null) {
-        renderer = gl.getParameter(debugInfo.UNMASKED_RENDERER_WEBGL);
-      }
-    }
-
-    // Mesa drivers may crash with more than 16 samplers and Firefox
-    // will actively refuse to create shaders with more than 16 samplers.
-    if (renderer && renderer.slice(0, 4).toUpperCase() === "MESA") {
-      maxSamplers = Math.min(16, maxSamplers);
-    }
-  } catch (exception) {
-    ErrorHandling.notify(exception as Error, {}, "warning");
-  }
-
-  return maxSamplers;
-}
-
-export function validateMinimumRequirements(specs: GpuSpecs): void {
-  if (specs.supportedTextureSize < 4096 || specs.maxTextureCount < 8) {
-    const msg =
-      "Your GPU is not able to render datasets in WEBKNOSSOS. The graphic card should support at least a texture size of 4096 and 8 textures.";
-    Toast.error(msg, {
-      sticky: true,
-    });
-    throw new Error(msg);
-  }
-}
 export type DataTextureSizeAndCount = {
   textureSize: number;
   textureCount: number;
